[feat] TI impersonation
This is a complete rewrite of the XS code since I could not get the
TI token. The code is a modified version of
github.com/lab52io/StopDefender. It runs on Windows 10. Windows 7
needs to be tested.
Also, removed dependency on WIN32 module and included the
WIN32::LoginName() funcntion from that code base. License
compatibility needs to be checked, but since WIN32 is under the
perl licence and perl leaves it open to use GPL or the Artistic
License, my interpretation is that this is fine.